Update suite.py [creating new 'effective_prompts' param]
It is useful to have a list of the actual prompts that we are using in the tests. In certain instances, these real prompts may differ from the 'prompts' attribute, so it is important to have a list of the effective prompts that we utilize. This prompt should be calculated once, as the order will be altered if we employ the 'shuffle' parameter.